src/services/produce_api/produceAPI/requirements.txt
argcomplete==1.11.1
asgiref==3.2.7
astroid==2.4.1
boto3==1.13.9
botocore==1.16.9
certifi==2020.4.5.1
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.1
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==7.1.2
pip-tools==5.1.2
placebo==0.9.0
pylint==2.5.2
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
tqdm==4.46.0
troposphere==2.6.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==0.16.1
wrapt==1.12.1
wsgi-request-logger==0.4.6
